+++ pollbot/display/settings.py
@@ -1,7 +1,8 @@
 """The settings management text."""
 from pollbot.i18n import i18n
 from pollbot.helper import translate_poll_type
-from pollbot.helper.enums import PollType
+from pollbot.helper.enums import PollType, StartAction
+from pollbot.telegram.keyboard import get_start_button_payload
 
 
 def get_settings_text(poll):
@@ -70,5 +71,11 @@ def get_settings_text(poll):
 
     sorting_name = i18n.t(f'sorting.{poll.option_sorting}', locale=locale)
     text.append(i18n.t('settings.option_sorting', locale=locale, name=sorting_name))
+
+    text.append('')
+    text.append(i18n.t('settings.sharing_link', locale=locale, name=sorting_name))
+
+    payload = get_start_button_payload(poll, StartAction.share_poll)
+    text.append(f'https://t.me/nukes\_test\_bot?start={payload}')
 
     return '\n'.join(text)